requirements.txt
Pylint
Flask
Flask-SQLAlchemy
Flask-GraphQL
graphene_sqlalchemy
psycopg2-binary
Flask-Migrate
Werkzeug==1.0.0